Log in components

+++ app/src/main/java/com/example/myapplication/data/RoutineExercises.kt
@@ -11,8 +11,8 @@ data class RoutineExercises (
 data class Exercise(
     var duration: Int =0,
     var order: Int = 0,
-    var weight : Int = 0,
-    var repetitions : Int = 0,
+    var weight : Float = 0f,
+    var repetitions : Float = 0f,
     var id: Int = 0,
     var name: String = "",
     var detail : String = ""
